+++ csrc/ops/alias.h
@@ -20,6 +20,9 @@
 
 namespace nvfuser {
 
+TORCH_CUDA_CU_API Val* set(Val*);
+TORCH_CUDA_CU_API TensorView* set(TensorView*);
+
 TORCH_CUDA_CU_API TensorView* view(TensorView* x, DataType dtype);
 
 TORCH_CUDA_CU_API TensorView* reshape(
