+++ test/test_gpu_indexing.cpp
@@ -12,7 +12,7 @@
 #include <inlining.h>
 #include <ir_all_nodes.h>
 #include <ir_builder.h>
-#include <ops/arith.h>
+#include <ops/all_ops.h>
 #include <scheduler/all_schedulers.h>
 
 #include <test/test_gpu_validator.h>
